LayoutTests/fast/frames/sandboxed-iframe-scripting-01.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
        var allowedExecuted = 0;
        var disallowedExecuted = 0;
        window.onload = function() {
            shouldBe("disallowedExecuted", "0");
            shouldBe("allowedExecuted", "1");

            isSuccessfullyParsed();
        };
    </script>
</head>
<body>
    <iframe sandbox="allow-same-origin allow-scripts"
            src="javascript: ++window.top.allowedExecuted;">
    </iframe>
    <script>
        description("Verify that sandboxed frames with sandbox='allow-scripts' can execute script.");
    </script>
</body>
</html>
