third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-scripting-01-expected.txt

Verify that sandboxed frames with sandbox='allow-scripts' can execute script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS disallowedExecuted is 0
PASS allowedExecuted is 1
PASS successfullyParsed is true

TEST COMPLETE
